Github for Mac Bug Report Test
To reproduce this issue, simply type any sufficiently long description
into the “Extended Description” box in Github for Mac. Create a branch
with one commit of this nature, and attempt to create a pull request.
The unwanted linebreaks will be evident in the automatically generated
pull request description. They should also be present when looking at
the commit log from the command line.